Fix `pytest` invocation
`pytest` was invoked with all source and extra paths, but it shouldn't.
Instead the path should be configured by the user in pyproject.toml.

If the user pass any positional arguments to the `pytest` sessions, then
those arguments will be forwarded verbatim to `pytest`.

Since `pytest` works with the default path `tests`, we leave that in the
extra paths in case the user is not defining it explicitly in the
configuration, but for more esoteric setups, the test paths should be
automatically added.

This commit also updates the default configuration for `api` to replace
`tests` with `pytests` in the extra paths (even when the user should
configure it explicitly, we want to remove `tests` as it might be for
something else) and also to replace `src` with `py` in the source paths.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>