RFC: Add a section on how to run nox.
It is a RFC because I'm not convinced the `requirements-nox.txt` is the
best approach, but if we use `pip install -e .[dev-nox]` or something
like that, then we are also installing all the normal package
dependencies, which are not really needed if all you want to do it just
run `nox`. If we don't do that, then users need a way to know which
version of `nox` they should install somehow (and any other dependencies
needed to run `nox`, like this package.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>