Avoid running sessions on sources for API repos
API repos are supposed to just ship python files generated from
protocol buffer files. Only a stub __init__.py should be included,
which doesn't really need to be checked.

Allowing checks on the sources for an API repos triggers a lot of issues
when the protocol buffers files are generated, as they don't pass any
checks. An alternative would be to exclude auto-generated files, but
it's not easy because each tool has its own exclusion mechanism, and
some don't even have any (hello darglint), so it is simpler to just skip
testing the stub __init__.py file.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>