Avoid the need to write a setup.py file for API repos
To do this we declare a distutils entry point in the pyproject.toml file
to automatically add the `compile_proto` command using the
`CompileProto` proto class.

Then, in the `grpc_tools` module we just add the newly declared command
as a `build` sub-command, so it is executed first. This make all other
subsequent sub-commands will see the generated sources as if they were
already there.

The only downside of this is now we don't have an easy way to configure
the command. The options are there, and can be passed via command-line
(like:

```sh
python -c 'import setuptools; setuptools.setup()' compile_proto \
    --proto-path=another/path
```

But there is no easy way to do it in a config file, or at least at the
moment I couldn't find a way to pass arbitrary options to setuptools
commands in `pyproject.toml`. If we need this, we could manually parse
the `pyproject.toml` file and look for options, just like
`setuptools_scm` does.

Signed-off-by: Leandro Lucarella <leandro.lucarella@frequenz.com>