Add dummy DCO job for the merge queue
The DCO GitHub app doesn't support merge queues yet, so to be able to
make DCO a required check for PRs, we need to provide a dummy DCO check
that will always pass for merge queues actions (if a PR is in the merge
we already know the DCO check passed anyway).

Note that we include the whole workflow in a jinja raw block to avoid
clashed between GitHub variable replacement and jinja variable
replacement, which overlap.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>