Unify `protobuf` generation configuration
A new module `protobuf` is added to handle the common `protobuf`
generation documentation which is used by both the `setuptools` plugin
to generate the Python files and the the `mkdocs` function to generate
the API documentation pages.

The `pyproject.toml` configuration key is changed from
`tool.frequenz-repo-config.setuptools.grpc_tools` to
`tool.frequenz-repo-config.protobuf` as it is more generic now.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>