Don't import modules into packages
Importing modules in package's __init__.py has the unintended effect of
indirectly importing everything the module imports too. This has a few
problems:

* Impacts performance, as many, probably unused, dependencies need to be
  imported
* Makes it impossible to use optional dependencies, as all dependencies
  will be imported anyway

Also not importing modules makes it easier to understand what is a
module and what is a symbol/object living in the package.

Because of this change, now `default` needs to be explicitly imported
in `noxfile.py`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>