Add golden testing for generated files
Testing the generated templates are correct is hard. For now we are
testing that the basics work by running `nox` on some generated projects
but this will not test bugs in documentation, comments, etc. Nor will
test GitHub Actions workflows or other advanced configuration.

By adding golden tests we at least ensure that once the results are
manually tested, we won't introduce unexpected regressions.

We considered using the pytest-golden plugin, but unfortunately it does
not support using whole files and directory trees as goldens, so we
built our own ad-hoc solution for now.

To update the golden files the test file needs to be temporarily updated
to set `UPDATE_GOLDEN` to `True` to replace the golden files with the
newly generated files by running `pytest` again, which is not great, but
it is good enough for now.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>