Add cookiecutter template variables description
The description of the variables will be printed before they are
prompted when generating a new project, and will be also included in the
documentation website.

To show the help before the project generation we need to use a very
dirty hack as cookiecutter doesn't support a hook that runs before the
variables are printed. So to do that we hijack a cookiecutter variable
itself ("Introduction") to display the help in the default text.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>