cookiecutter: Add workflow to check for release notes update
This workflow will check PRs to see if a change was done in the `src/`
directory, and if so, it will fail if the `RELEASE_NOTES.md` wasn't also
updated.

Users can override this by assigning the label `cmd:skip-release-notes`
to the PR for changes that don't really need a release notes update.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>