Add support for linting code examples in docstrings
This commit adds a module with an utility function to be able to easily
collect and lint code examples in docstrings.

It also add an optional dependency to easily pull the dependencies
needed to do this linting and ignores any optional dependency starting
with `extra-` in the tests checking if new repo types were added, as
optional dependencies starting with `extra-` are not really repository
types.

This is based on the work on the SDK:
https://github.com/frequenz-floss/frequenz-sdk-python/pull/384

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>