Don't ship development files in source distribution
There are a lot of files that are only used for development purposes
and should not be distributed in the source distribution package, as
this package should only include the files necessary to build the wheel.

We still ship basic documentation (like the README, release notes and
license as they are very lightweight and carry important information).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>