cookiecutter: Add a `pytest` hook to lint docstring examples
Examples found in code *docstrings* in the `src/` directory will now be
collected and checked using `pylint`. This is done via the file
`src/conftest.py`, which hooks into `pytest`.

We also need to update the regex to use the current development version
of the package in the tests because now we also use the
`extra-lint-examples` optional dependency.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>