Replace `darglint` with `pydoclint` via `flake8`
The `darglint` project is not maintained anymore, and even when there is
a fork that is trying to keep it alive, it is still extremely slow.

The new project `pydoclint` was created recently but advancing rapidly,
and already in better shape than `darglint` and way faster (0.2s vs 20s
for the SDK, 100x improvement).

To be able to add ignore comments in the code, `pydoclint` needs to be
run via `flake8`. Because we need to run `flake8` already, and
`pydocstyle` also can run via `flake8`, we are merging both in one
`flake8` call, so we only need to read the files once.

As a side effect, we are also enabling other base `flake8` checks. There
is probably some overlap with `pylint`, but `flake8` is very fast
anyway, so it shouldn't be noticed (for the SDK, `flake8` basic checks
+ `pydocstyle` + `pydoclint` runs in 0.6s and `pylint` alone runs in
15s). At some point we might want to disable the duplicated checks in
`pylint` to see if it speeds up `pylint`.

This commit also renames the `dev-docstrings` optional dependency to
`dev-flake8` because now is not checking docstrings exclusively.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>