Group dependabot dependency updates
We group production and development ("optional" in the context of
pyproject.toml) dependency updates when they are patch and minor
updates,so we end up with less PRs being generated.

Major updates are still managed, but they'll create one PR per
dependency, as major updates are expected to be breaking, it is better
to manage them individually.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>