ci: Add env and pip debug information
When debugging different behaviour between the GitHub workflow run and
a local run it is very useful to know exactly which dependency versions
are installed.

To do this we also add the `nox` session as a matrix dimention for the
`nox` job, so it is easier to print the generated `nox` venv for the
different sessions. This has the nice effect that nox max and min tests
will run in parallel.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>