Split `mypy` check for the source package and dev files
`mypy` tends to have issues with checking source files inside a `src/`
directory. There are tricks[1] but it seems like the only way to do it
is to run `mypy` on `.` using `MYPYPATH=src`, which will still check the
whole `.` directory (include whole virtualenvs if there are any), which
is not what we want. For now we've been using `-p` to check for packages
instead of files, which seemed to have done the trick, but we need to
pass directories that don't really have a package structure (like
benchmarks or examples or tests) as packages, which is also not great.

This commit tries a different approach, we check the source code as a
package separately, and then check the development files (tests,
examples, etc.), which are passed as simple paths. This also opens up
the door to using more relaxed rules for these development files if
needed in the future.

To achieve this we also move the `packages` specification to the
`pyproject.toml` file, so now `mypy` can be called without any arguments
to check the source code package, making it easier to call it without
relaying in `nox`.

[1] https://mypy.readthedocs.io/en/stable/running_mypy.html#mapping-file-paths-to-modules

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>