Add `mypy` no-incremental option as a comment
This option avoids using a cache. It is slower but prevents some issues
with `mypy` giving different results on different runs (typically
complaining about some `type: ignore[issue]` not being used but when
removing it complaining about an `issue` again.

It is included so it has more visibility to users in the hope that they
can spend less time debugging if they hit this `mypy` issue.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>