Add tests for non-native architectures
This change introduces a new job to the CI to test the SDK on non-native
architectures. Github CI ATM provides amd64 runners by default, so the
new job targets testing on arch64 for now.

The tests are run in containers running on the target platform(s). The
container definitions are supposed to exist in the directory
.github/containers/nox-cross-arch. A definition for an
arm64-ubuntu20.04-python3.11 container can be found there.

Since the tests take quite a while to run, given that they run on
emulations, this step runs only on the `merge_group` event.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>