Move cookiecutter filters to a Jinja2 extension
By using a full fledged Jinja2 extension we can not only add filters but
also variables. This also simplifies the `cookiecutter.json` file, as we
don't need to enumerate every single filter we want to use. This also
simplifies the template update process as when new fitlers are added
there is no need to update the cookiecutter replay file, also minimizing
chances of weird erros when that happens.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>