Use a random day of the week for Dependabot updates
To avoid dependabot updates flooding users notification with updates to
all projects at the same time, we use a random day of the week for
running the updates. For this we add a `random_weekday` global variable
to the Jinja2 environment that will be available in all templates. The
random value is calculated once, so it will be the same each time is
used during the same run.

When running inside `pytest` we always use Monday to make the tests
reproduceable and we also use Monday for this project.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>