Don't run cookiecutter tests in arm64
We don't really create project in arm64 yet, so it's not strictly
necessary to have it tested, and test in arm64 take too long as they
need to run in a emulator.

If at some point arm64 is popular enough so we want to make sure new
projects can be started in it, there will probably be more accessible
native arm64 GitHub Actions runners.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>