Add a command-line tool to get the current `mike` version
For now this tool is designed to be used in GitHub Actions workflows,
but is is possible to use it in other environments as well doing some
extra configuration.

To use this tool call it via python:

`python -m frequenz.repo.config.cli.version.mike.info`

In the future the idea will be to have a command with sub-commands, like
`repo-config version mike info`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>