Add method to find the next branch introducing breaking changes
If there is a last tag, use that as a base, otherwise use the current
branch as a base. If none is available, return `None`. If there is
a base, the major is incremented by one and the minor is set to `None`,
unless the major is 0 (an "initial development version" for semver), in
which case the major is set to 0 and the minor is incremented by one, as
the next minor could be a breaking change. Technically semver allows
breaking changes in patches for major version 0, but we assume patches
maintain backwards compatibility.

For example:

- `v0.x.x` -> `v0.2.x` (if the last tag is `v0.1.1` for example)
- `v0.2.x` -> `v0.3.x`
- `v1.x.x` -> `v2.x.x`
- `v1.0.x` -> `v2.x.x`
- `v0.0.1` -> `v0.1.x`
- `v0.3.1-pre.1+build.3` -> `v0.4.x`
- `v1.0.0` -> `v2.x.x`
- `v1.1.0` -> `v2.x.x`

This is useful to use as a maximum dependency version for a package.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>