Upgrade `mike` to 2.0.0
This commit upgrades `mike` to 2.0.0. This new version uses symlinks
for aliases by default, but they are not supported by GitHub Pages, so
we need to use the `redirect` alias type instead explicitly.

We use the new `alias_type` option in the `mike` plugin configuration
in the `mkdocs.yml` file to set the alias type to `redirect`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>