Don't fail if the `mike` version can't be determined
If the `mike` version can't be determined, just emit a warning and exit
succesfully. This is useful for the CI, where the `mike` version can't
be determined because the branch name doesn't follow the guidelines (for
example in PRs).

The CI will now emit a warning and skip the publishing of the
documentation website.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>