Add instructions to configure the GitHub repository
This is based on the instructions in the Wiki:
https://github.com/frequenz-floss/frequenz-repo-config-python/wiki/Configuring-a-new-GitHub-repository

Only a few adjustments are made to fix a few out of date data and to
make it fit better to the current existing documentation.

The instructions to configure the `RELEASE_NOTES.md` update check is
removed because now it's spread in the rest of the configuration (the
adding of the label is moved to the Labels section, and the check is
included in the rulesets now).

After this commit is merged they will be superseded by the instructions
in the docs.

To build the links to the rulesets, we add a new mkdocs macro variable
to get the current default branch, which is only used when there is no
GitHub information about the current version (mainly when running
locally).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>