Properly translate paths from the python package name
When the `package_name` is changed from the default, we need to make
sure we adapt the paths in the generated project to match the new
package name, otherwise those packages will not be found.

With this general approach, we can also remove the special case for
`lib` projects, as it is now handled by the general case.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>