Treat `pytest` warnings as errors
Warnings are usually a sign of something that can be improved or even
wrong, like tasks not being properly awaited or cancelled, so it's a
good idea to treat them as errors to avoid them getting unnoticed.

We still treat deprecations as warnings, as when testing with the
`pytest_min` session is normal to get deprecation warnings as we are
using old versions of dependencies.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>