Fix setuptools dependency
It looks like the deependency got a bit mixed up at some point, and
`types-setuptools` dependency started getting widenned and the real
optional dependency (for `api`) stopped being widenned.

We'll probably need to keep an eye on dependabot to see if this was done
by dependabot, but even if it is not, it is failing to widden the
`setuptools` dependency for some reason.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>