Use the new private ARM runner for private repositories
We used to use the same name for our private ARM runner as the public
one from GitHub, so we could easily re-use the same workflow for private
and public repos. This doesn't work anymore because GitHub now randomly
picks any runner with that name, and if it is a public one, then it will
refuse to run for private repos. Because of this we now use the suffix
`-arm64` for our private ARM runner, and we target the right runner
based on the `github.repository_visibility` context variable.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>