Do not convert array of paths to a string
When loading protobuf include paths from the pyproject.toml file, the
paths are received as an array/list. If we convert them to
a comma-separated string there is the risk of a path containing commas
to be split incorrectly, leading to compilation errors.

Using the original array/list of paths instead of a comma-separated
string is safer and more flexible.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>